Flask/requirements.txt
absl-py==1.0.0
aiofiles==23.2.1
altair==5.3.0
altgraph==0.17.4
annotated-types==0.6.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
backports.tarfile==1.0.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.6.2
bottle==0.12.21
cachetools==5.1.0
Calculator==1.1.0
certifi==2024.2.2
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.8
docker==7.0.0
editables==0.5
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.1.1
export==0.2.0
fastapi==0.110.1
fastjsonschema==2.16.2
ffmpy==0.3.2
filelock==3.12.0
Flask==2.3.2
Flask-Login==0.6.3
flask-ngrok==0.0.25
flatbuffers==23.5.26
fonttools==4.39.0
fsspec==2023.12.2
gast==0.4.0
gdown==4.6.3
gitdb==4.0.10
GitPython==3.1.31
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-generativeai==0.5.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
googletrans==4.0.0rc1
gradio==4.26.0
gradio_client==0.15.1
grpcio==1.62.1
grpcio-status==1.62.1
gTTS==2.4.0
guizero==1.3.0
h11==0.9.0
h2==3.2.0
h5py==3.7.0
hatch==1.9.4
hatchling==1.21.1
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.19.4
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
ikpy==3.0.1
importlib-metadata==4.11.4
importlib-resources==5.12.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
itsdangerous==2.1.2
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.10
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter-server==1.21.0
jupyter_client==7.4.2
jupyterlab==3.4.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.16.0
kaggle==1.6.12
kagglehub==0.2.2
keras==2.9.0
Keras-Preprocessing==1.1.2
keyring==25.1.0
kiwisolver==1.4.4
libclang==14.0.1
lxml==4.9.1
Markdown==3.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.9
mistune==0.8.4
more-itertools==10.2.0
mpmath==1.2.1
mysql-connector-python==8.0.31
nbclassic==0.4.5
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.1
notebook-shim==0.1.0
numpy==1.22.4
oauthlib==3.2.0
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.10.0
outcome==1.2.0
packaging==24.0
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
pefile==2023.2.7
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.5.1
pluggy==1.4.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
pygame==2.1.2
Pygments==2.13.0
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.4
pyparsing==3.0.9
pyperclip==1.8.2
pypot==5.0.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2022.4
pywin32==304
pywin32-ctypes==0.2.2
pywinpty==2.0.8
PyYAML==6.0
pyzmq==24.0.1
regex==2023.10.3
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.7.1
rsa==4.8
ruff==0.3.7
safetensors==0.4.1
scikit-learn==1.3.0
scipy==1.8.1
seaborn==0.12.2
selenium==4.8.2
semantic-version==2.10.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.1
starlette==0.37.2
sympy==1.10.1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.16.0
text-unidecode==1.3
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.15.0
tomli==2.0.1
tomli_w==1.0.0
tomlkit==0.12.0
toolz==0.12.1
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tornado==6.2
tqdm==4.65.0
traitlets==5.4.0
transformers==4.36.2
trio==0.22.0
trio-websocket==0.9.2
trove-classifiers==2024.4.10
typer==0.12.3
typing_extensions==4.11.0
tzdata==2022.5
ultralytics==8.2.2
uritemplate==4.1.1
urllib3==1.26.9
userpath==1.9.2
uvicorn==0.29.0
virtualenv==20.23.0
voila==0.3.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
websockets==10.3
Werkzeug==2.3.4
wget==3.2
widgetsnbextension==4.0.7
wrapt==1.14.1
wsproto==1.2.0
zipp==3.8.0
zstandard==0.22.0
